safari-600.7-branch/LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/test_timing_xserver_redirect-expected.txt

Description

This test validates the values of the window.performance.redirectCount and the window.performance.timing.redirectStart/End times for a cross-origin server side redirect navigation.




FAIL Starting document.location.hostname is correct (127.0.0.1:8000) assert_equals: Starting document.location.hostname is correct (127.0.0.1:8000) expected "127.0.0.1:8000" but got "127.0.0.1"
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS timing.navigation.type is TYPE_NAVIGATE 
PASS navigation.redirectCount == 0 on a cross-origin server redirected navigation 
PASS window.performance.timing.navigationStart > 0 
PASS timing.redirectStart == 0 on a server redirected navigation from another domain 
PASS timing.redirectEnd == 0 on a server redirected navigation from another domain 
